requirements.txt
tensorflow
fastapi
uvicorn[standard]
pillow
streamlit
numpy
python-multipart
